% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter_expr.R
\name{create_filter_expr}
\alias{create_filter_expr}
\title{Create filter expression for a single filter}
\usage{
create_filter_expr(x, var, val, include_na = TRUE)
}
\arguments{
\item{x}{a numeric/character vector}

\item{var}{name of a filter variable}

\item{val}{values from input controls (range slider / drop-down menu)}

\item{include_na}{a logical value indicating whether missing values (NA)
should be included}
}
\description{
Create filter expression for a single filter
}
