% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter_server.R
\name{data_filter_server}
\alias{data_filter_server}
\title{Server function of the data filter module}
\usage{
data_filter_server(id, data)
}
\arguments{
\item{id}{An ID string of the module's server function.}

\item{data}{A reactive data for dynamic filtering.}
}
\description{
Server function of the data filter module
}
