% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filter_ui.R
\name{data_filter_ui}
\alias{data_filter_ui}
\title{UI function of the data filter module}
\usage{
data_filter_ui(id)
}
\arguments{
\item{id}{An ID string of the module's UI function.}
}
\description{
UI function of the data filter module
}
